refactor(search): extract DisplayRow into display_row.rs, drop backend.rs size exception
backend.rs was 1067 LOC and carried a PERMANENT file_size_exceptions
entry. Per workspace policy (decompose, don't suppress), move the
self-contained DisplayRow type — struct + inherent impl + Default +
uffs_format::FormatRow impl — into a new sibling module display_row.rs
(289 LOC). backend.rs drops to 784 LOC, under the 800 ceiling.

DisplayRow is re-exported (`pub use super::display_row::DisplayRow;`)
so the single-import convention downstream relies on
(uffs_core::search::backend::DisplayRow) is unchanged — public API and
behavior preserved.

Removes the backend.rs entry from scripts/ci/file_size_exceptions.txt.
